requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Werkzeug
SQLAlchemy
google-generativeai
python-dotenv
Pillow

PyPDF2
python-docx